requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==12.0.1
Flask==2.1.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
mixer==7.2.2
numpy==1.22.3
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
